Refactor workflow analyzer to support cross-file dependency tracking
