+++ package.json
@@ -23,8 +23,7 @@
     "clean": "turbo run clean",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
     "format:check": "prettier --check \"**/*.{js,jsx,ts,tsx,json,md}\"",
-    "analyze:workflows": "tsx scripts/analyze-workflows.ts",
-    "analyze:workflows:enhanced": "tsx scripts/analyze-workflows-enhanced.ts"
+    "analyze:workflows": "tsx scripts/analyze-workflows.ts"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "catalog:",
